bug testing

+++ hw3/model.c
@@ -179,7 +179,7 @@ int GetCandy(BoardPtr gameboard, int idx) {
 
     if(gameboard->array_ptr != NULL) { //&& idx < GetSize(gameboard->array_ptr)) {
         GetElement(gameboard->array_ptr, idx, ptr);
-        val = *ptr;
+        val = (int)*ptr;
         printf("val1: %d\n", (int)**ptr);
         printf("val2: %d\n", (int)*ptr);
     } else {
